move tests to general storage behavior tests and remove unused imports
