+++ tests/storage/cdi_import/test_import_http.py
@@ -15,7 +15,6 @@ from tests.os_params import RHEL_LATEST
 from tests.storage.cdi_import.utils import (
     get_importer_pod_node,
     wait_dv_and_get_importer,
-    wait_for_pvc_recreate,
 )
 from tests.storage.constants import (
     ALPINE_QCOW2_IMG,
@@ -31,9 +30,7 @@ from tests.storage.utils import (
     get_importer_pod,
     wait_for_importer_container_message,
 )
-from utilities import console
 from utilities.constants import (
-    OS_FLAVOR_FEDORA,
     OS_FLAVOR_RHEL,
     QUARANTINED,
     TIMEOUT_1MIN,
@@ -45,11 +42,8 @@ from utilities.infra import get_node_selector_dict
 from utilities.ssp import validate_os_info_vmi_vs_windows_os
 from utilities.storage import (
     ErrorMsg,
-    create_dummy_first_consumer_pod,
     create_dv,
     create_vm_from_dv,
-    get_dv_size_from_datasource,
-    sc_volume_binding_mode_is_wffc,
 )
 from utilities.virt import running_vm
 
@@ -64,32 +58,6 @@ TAR_IMG = "archive.tar"
 DEFAULT_DV_SIZE = Images.Alpine.DEFAULT_DV_SIZE
 SMALL_DV_SIZE = "200Mi"
 LATEST_WINDOWS_OS_DICT = py_config.get("latest_windows_os_dict", {})
-
-
-@pytest.mark.sno
-@pytest.mark.polarion("CNV-675")
-def test_pvc_recreates_after_deletion(namespace, storage_class_name_scope_function, fedora_data_source_scope_module):
-    with create_dv(
-        dv_name=f"cnv-675-{storage_class_name_scope_function}",
-        namespace=namespace.name,
-        storage_class=storage_class_name_scope_function,
-        size=get_dv_size_from_datasource(fedora_data_source_scope_module),
-        client=namespace.client,
-        source_ref={
-            "kind": fedora_data_source_scope_module.kind,
-            "name": fedora_data_source_scope_module.name,
-            "namespace": fedora_data_source_scope_module.namespace,
-        },
-    ) as dv:
-        dv.wait_for_dv_success(timeout=TIMEOUT_5MIN)
-        pvc = dv.pvc
-        pvc_original_timestamp = pvc.instance.metadata.creationTimestamp
-        pvc.delete()
-        wait_for_pvc_recreate(pvc=pvc, pvc_original_timestamp=pvc_original_timestamp)
-        storage_class = storage_class_name_scope_function
-        if sc_volume_binding_mode_is_wffc(sc=storage_class, client=namespace.client):
-            create_dummy_first_consumer_pod(pvc=pvc)
-        dv.wait_for_dv_success()
 
 
 @pytest.mark.xfail(
@@ -370,38 +338,6 @@ def test_successful_concurrent_blank_disk_import(
 @pytest.mark.s390x
 def test_blank_disk_import_validate_status(data_volume_multi_storage_scope_function):
     data_volume_multi_storage_scope_function.wait_for_dv_success(timeout=TIMEOUT_5MIN)
-
-
-@pytest.mark.polarion("CNV-3065")
-@pytest.mark.sno
-def test_disk_falloc(
-    storage_class_name_scope_function, unprivileged_client, fedora_data_source_scope_module, namespace
-):
-    size = get_dv_size_from_datasource(data_source=fedora_data_source_scope_module)
-    with create_dv(
-        client=unprivileged_client,
-        dv_name=f"cnv-3065-{storage_class_name_scope_function}",
-        namespace=namespace.name,
-        source_ref={
-            "kind": fedora_data_source_scope_module.kind,
-            "name": fedora_data_source_scope_module.name,
-            "namespace": fedora_data_source_scope_module.namespace,
-        },
-        size=size,
-        storage_class=storage_class_name_scope_function,
-    ) as dv:
-        dv.wait_for_dv_success(timeout=TIMEOUT_5MIN)
-        with create_vm_from_dv(
-            vm_name="cnv-3065-vm",
-            client=unprivileged_client,
-            dv=dv,
-            os_flavor=OS_FLAVOR_FEDORA,
-            memory_guest=Images.Fedora.DEFAULT_MEMORY_SIZE,
-        ) as vm_dv:
-            with console.Console(vm=vm_dv) as vm_console:
-                LOGGER.info(f"Fill disk space with size {size}")
-                vm_console.sendline(f"fallocate -l {size} test-file")
-                vm_console.expect("No space left on device", timeout=TIMEOUT_1MIN)
 
 
 @pytest.mark.destructive
